Introduce common dvv modification service integration test base
