+++ service/src/integrationTest/kotlin/fi/espoo/evaka/dvv/DvvModificationServiceClientIntegrationTest.kt
@@ -3,12 +3,11 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 package fi.espoo.evaka.dvv
 
-import fi.espoo.evaka.FullApplicationTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
 
-class DvvModificationServiceClientIntegrationTest : FullApplicationTest() {
+class DvvModificationServiceClientIntegrationTest : DvvModificationServiceIntegrationTestBase() {
 
     @Test
     fun `get modification token for today`() {
