+++ service/src/integrationTest/kotlin/fi/espoo/evaka/dvv/DvvModificationServiceIntegrationTest.kt
@@ -4,18 +4,35 @@
 package fi.espoo.evaka.dvv
 
 import deleteDvvModificationToken
-import fi.espoo.evaka.FullApplicationTest
+import fi.espoo.evaka.resetDatabase
+import fi.espoo.evaka.shared.db.handle
 import getDvvModificationToken
 import getNextDvvModificationToken
 import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import storeDvvModificationToken
 
-class DvvModificationServiceIntegrationTest : FullApplicationTest() {
+class DvvModificationServiceIntegrationTest : DvvModificationServiceIntegrationTestBase() {
+
+    @BeforeEach
+    private fun beforeEach() {
+        jdbi.handle { h ->
+            resetDatabase(h)
+            storeDvvModificationToken(jdbi, "100", "101", 0, 0)
+        }
+    }
+
+    @BeforeEach
+    private fun afterEach() {
+        jdbi.handle { h ->
+            deleteDvvModificationToken(jdbi, "100")
+        }
+    }
 
     @Test
-    fun `get modification token for today`() {
-        storeDvvModificationToken(jdbi, "100", "101", 0, 0)
+    fun `get modification token for today`() = jdbi.handle { h ->
+
         assertEquals("101", getNextDvvModificationToken(jdbi))
         val response = dvvModificationsService.getDvvModifications(listOf("nimenmuutos"))
         assertEquals(1, response.size)
@@ -25,7 +42,12 @@ class DvvModificationServiceIntegrationTest : FullApplicationTest() {
         assertEquals("102", createdDvvModificationToken.nextToken)
         assertEquals(1, createdDvvModificationToken.ssnsSent)
         assertEquals(1, createdDvvModificationToken.modificationsReceived)
-        deleteDvvModificationToken(jdbi, "100")
+
         deleteDvvModificationToken(jdbi, "101")
+    }
+
+    @Test
+    fun `person date of death`() {
+        dvvModificationsService.updatePersonsFromDvv(listOf("kuollut"))
     }
 }
